Tried to finish the initial implementation of vector indexes by adding tests and making sure that the functionality works with both mock and local embeddings.
