+++ .vscode/settings.json
@@ -1,4 +1,5 @@
 {
+    "editor.parameterHints.enabled": false,
     "python.testing.pytestArgs": [
         "tests"
     ],
