+++ mypy.ini
@@ -8,7 +8,7 @@ check_untyped_defs = True
 disallow_untyped_decorators = True
 no_implicit_optional = True
 strict_optional = True
-mypy_path = stubs
+mypy_path = src/stubs
 
 [mypy.plugins.numpy.*]
 follow_imports = skip
