+++ pyproject.toml
@@ -17,6 +17,9 @@ dependencies = [
     "pytest-cov>=6.1.1",
     "sentence-transformers>=4.1.0",
     "openai>=1.79.0",
+    "sqlite-utils>=3.38",
+    "setuptools>=80.7.1",
+    "types-setuptools>=80.8.0.20250521",
 ]
 
 [project.optional-dependencies]
